dqe/requirements.txt
geojson-pydantic==0.2.2
jinja2==2.11.3
cachetools==4.2.1
pystac==1.0.0-rc.2
stac_vrt==1.0.3
stac-pydantic==2.0.0
planetary-computer==0.3.0-rc.0

rasterio==1.2.6
titiler.core==0.3.*
titiler.mosaic==0.3.*
titiler.application==0.3.*
rio-tiler==2.1.*
matplotlib==3.4.*

uvicorn[standard]==0.13.3
uvloop==0.14.0
gunicorn==20.1.0

azure-identity==1.5.0
azure-storage-blob==12.6.0

opencensus-ext-azure==1.0.8
opencensus-ext-logging==0.1.0
